tango/src/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2020.6.20
cffi==1.16.0
chardet==3.0.4
colorama==0.4.3
crispy-bootstrap4==2024.1
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.0
Django==5.0.2
django-allauth==0.61.1
django-bootstrap4==24.1
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-js-asset==2.2.0
django-markdownx==4.0.7
django-tinymce==3.7.1
et_xmlfile==1.0.1
filelock==3.0.12
idna==2.10
jdcal==1.4.1
Markdown==3.5.2
oauthlib==3.2.2
openpyxl==3.0.4
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pygame==2.5.2
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.1
six==1.15.0
soupsieve==2.5
sqlparse==0.3.1
tzdata==2024.1
urllib3==1.25.10
virtualenv==20.0.25
virtualenvwrapper-win==1.2.6
